pyproject.toml
[project]
name = "cognito-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "authlib>=1.4.0",
    "boto3>=1.35.92",
    "fastapi>=0.115.6",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "python-dotenv>=1.0.1",
    "starlette>=0.41.3",
    "uvicorn>=0.34.0",
]
